+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Killer sudoku solver".
 - Tests for "IP mask calculating".
 - Tests for "Domino puzzle".
+- Tests for "Goldbach’s conjecture".
 
 ## [1.16.0] - 2022-12-31
 ### Added